Work_Job_Tracker_2.0/requirements.txt
mysql-connector-python==9.5.0
python-dotenv==1.2.1
tkcalendar==1.6.1